test(workers): keep native provider seam checks focused
